requirements.txt
boto3==1.26.137
python-dotenv==1.0.0
requests==2.28.2
pandas==2.1.3
watchtower==3.0.1  # For CloudWatch logging
python-json-logger==2.0.7  # For structured JSON logging